modals

+++ servidor/public/js/index.js
@@ -18,7 +18,11 @@ function validateEmail() {
     const test = regexEmail.test(email);
     if (test && email!="") {
         return true
+    }  else if (email == "") {
+        $(".errorEmail").text("Este campo es obligatorio")
+        showError(errorEmail)
     } else {
+        $(".errorEmail").text("Email invalido")
         showError(errorEmail)
     }
 }
@@ -28,7 +32,11 @@ function validateName() {
     const errorName = $(".errorName")
     if (val.length < 30 && val!="") {
         return true
+    } else if (val == "") {
+        $(".errorName").text("Este campo es obligatorio")
+        showError(errorName)
     } else {
+        $(".errorName").text("El nombre tiene que ser menor a 30 caracteres")
         showError(errorName)
     }
 }
@@ -38,7 +46,11 @@ function validateSurname() {
     const errorSurname = $(".errorSurname")
     if (val.length < 30 && val!="") {
         return true
+    } else if (val == "") {
+        $(".errorSurname").text("Este campo es obligatorio")
+        showError(errorSurname)
     } else {
+        $(".errorSurname").text("El apellido tiene que ser menor a 30 caracteres")
         showError(errorSurname)
     }
 }
@@ -49,10 +61,15 @@ function validateNumbers() {
     test = range.test(phoneInput.val())
     if (test && phoneInput.val()!="") {
         return true
+    } else if (phoneInput.val() == "") {
+        $(".errorPhone").text("Este campo es obligatorio")
+        showError(errorPhone)
     } else {
+        $(".errorPhone").text("El telefono solo pueden ser numeros")
         showError(errorPhone)
     }
 }
+
 
 // nuevo usuario
 $("#newUser").on("click", function() {
@@ -76,11 +93,23 @@ $("#saveForm").on("click",  function(e) {
                 phone: phoneInput.val(),
                 email: emailInput.val()
             },
-            success: window.location.href = "http://localhost:3000/users"
+            success: function() {
+                // modal
+                Swal({
+                    type: 'success',
+                    heightAuto: false,
+                    title: 'Usuario creado',
+                    showConfirmButton: false,
+                    timer: 1500
+                })
+                setTimeout(function() {
+                    window.location.href = "http://localhost:3000/users"
+                }, 2000)
+            }
         })
     }
-    // Swal('Creaste un usuario nuevo')
 })
+
 
 
 // appendear users
@@ -94,9 +123,28 @@ $.ajax("http://localhost:3000/api/users").done(function(data) {
 // borrar user
 $(document).on("click", ".erase-but", function() {
     let id = $(this).parent().parent().attr("id")
-    $.ajax("http://localhost:3000/api/users/"+id, {
-        method: "DELETE", 
-        success: window.location.reload()
+    // modal
+    Swal({
+        title: 'Seguro queres borrar este user?',
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si'
+    }).then((result) => {
+        if (result.value) {
+            $.ajax("http://localhost:3000/api/users/"+id, {
+                method: "DELETE", 
+                success: setTimeout(function() {
+                    window.location.reload()
+                }, 1500) 
+            })
+            Swal({
+                title: 'Usuario borrado',
+                type: 'success',
+                showConfirmButton: false,
+            })
+        }
     })
 })
